requirements.txt
# Backend Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==3.0.2
websockets==15.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.20

# OpenAI API
openai==1.108.1
tiktoken==0.9.0

# Google Drive Integration
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# File Processing
python-docx==1.1.2
PyPDF2==3.0.1
PyYAML==6.0.2

# Utilities
aiofiles==24.1.0
requests==2.32.4

# Optional: For academic humanizer (if used)
# nltk==3.9.1
# spacy
# sentence-transformers==2.5.1

